commit before pull
