+++ client/src/components/Signup.js
@@ -8,96 +8,33 @@ export const Signup = () => {
     const errRef = useRef();
 
     const [user, setUser] = useState('');
+    const [validName, setValidName] = useState(false);
+    const [userFocus, setUserFocus] = useState(false);
+
     const [pwd, setPwd] = useState('');
-    const [errMsg, setErrMsg] = useState('');
+    const [validPwd, setValidPwd] = useState(false);
+    const [pwdFocus, setPwdFocus] = useState(false);
+
+    const [matchPwd, setMatchPwd] = useState("");
+    const [validMatch, setValidMatch] = useState(false);
+    const [matchFocus, setMatchFocus] = useState(false);
+  
+    const [errMsg, setErrMsg] = useState("");
     const [success, setSuccess] = useState(false);
 
-    const [email] = useState('');
-    const [lastName] = useState('');
+    useEffect(() => {
+        useRef.current.focus();
+    }, []);
 
     useEffect(() => {
-        userRef.current.focus();
-    }, [])
-
-    useEffect(() => {
-        setErrMsg('');
-    }, [user, pwd])
-
-    const handleSubmit = async (e) => {
-        e.preventDefault();
-        console.log(user, pwd);
-        setUser('')
-        setPwd('')
-        setSuccess(true);
-    }
-
+        setValidName(USER_REGEX.test(user));
+    }, [user]);
 
     return (
-        <>
-            <div>
-            <p ref={errRef} className={errMsg ? "errmsg" : "offscreen"} aria-live="assertive"></p>
-            
-            <h1>Sign Up</h1>
-
-            <form onSubmit={handleSubmit}>
-                {/* email */}
-                <label htmlFor="email">Email:</label>
-                <input 
-                    type="text" 
-                    id="email"
-                    ref={userRef}
-                    autoComplete="off"
-                    onChange={(e) => setUser(e.target.value)}
-                    value={email}
-                    required
-                />
-                < br/>
-                {/* first name */}
-                <label htmlFor="firstName">First Name:</label>
-                <input 
-                    type="text" 
-                    id="firstName"
-                    ref={userRef}
-                    autoComplete="off"
-                    onChange={(e) => setUser(e.target.value)}
-                    value={user}
-                    required
-                />
-                {/* last name */}
-                 <label htmlFor="email">Last Name:</label>
-                <input 
-                    type="text" 
-                    id="lastName"
-                    ref={userRef}
-                    autoComplete="off"
-                    onChange={(e) => setUser(e.target.value)}
-                    value={lastName}
-                    required
-                />
-                {/* password */}
-                <label htmlFor="password">Password:</label>
-                <input 
-                    type="password" 
-                    id="password"
-                    ref={userRef}
-                    onChange={(e) => setPwd(e.target.value)}
-                    value={pwd}
-                    required
-                />
-                < br />
-                <button>Sign In</button>
-
-                <p>
-                    Already have an account? < br/>
-                    <span>
-                        { /*insert route */}
-                        <a href="#">Sign In</a>
-                    </span>
-                </p>
-            </form>
-
+        <div>
+            sample text
         </div>
-        </>
-        
     )
 }
+
+export default Signup
