+++ package.json
@@ -16,6 +16,7 @@
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "express-session": "^1.17.3",
+    "nodemon": "^2.0.20",
     "pg": "^8.8.0"
   }
 }
